agents/industry_research_agent/industry_research.py
import yaml

class IndustryResearchAgent:
    def __init__(self, config_path):
        with open(config_path, 'r') as file:
            self.config = yaml.safe_load(file)

    def fetch_industry_info(self, company_name):
        return {"industry": "Retail", "offerings": "E-commerce, Supply Chain Management"}

    def segment_industry(self, industry_data):
        return industry_data.get("industry"), industry_data.get("offerings")
