requirements.txt

pyyaml
requests
